+++ data_raw/data_raw.R
@@ -5,6 +5,8 @@
 # Parameters ----
 `%>%` <- magrittr::`%>%`
 api_key <- Sys.getenv("eia_key")
+
+# US residential consumption ----
 category_details <- tsAPI::eia_query(api_key = api_key,category_id = 480585)
 series_details <- category_details$category$childseries
 series_details$f
@@ -12,7 +14,7 @@ monthly <- series_details$series_id[which(series_details$f == "M")]
 monthly
 
 
-us_res_gas <- lapply(1:length(monthly), function(i){
+us_residential <- lapply(1:length(monthly), function(i){
   id <- monthly[i]
   series_name <- series_details$name[which(series_details$series_id == id)]
   state <- substr(series_name, 1, regexpr("Natural Gas", text = series_name) - 2)
@@ -30,6 +32,37 @@ us_res_gas <- lapply(1:length(monthly), function(i){
     dplyr::mutate(date = lubridate::ymd(paste(date, "01", sep = ""))) %>%
     dplyr::mutate(state = state) %>%
     dplyr::select(date, state, y)
+
+}) %>%
+  dplyr::bind_rows()
+
+usethis::use_data(us_residential, overwrite = TRUE)
+write.csv(us_residential, "csv/us_residential.csv", row.names = FALSE)
+
+# US total consumption ----
+category_details <- tsAPI::eia_query(api_key = api_key,category_id = 480324)
+series_details <- category_details$category$childseries
+series_details$f
+annual <- series_details$series_id[which(series_details$f == "A")]
+annual
+
+us_total <- lapply(1:length(annual), function(i){
+  id <- annual[i]
+  series_name <- series_details$name[which(series_details$series_id == id)]
+  state <- substr(series_name, 1, regexpr("Natural Gas", text = series_name) - 2)
+  url <- paste("curl --location --request GET 'http://api.eia.gov/series/?api_key=",
+               api_key,
+               "&series_id=",
+               id,
+               "' | jq -r '.series[].data[] | [.[0], .[1]] | @tsv'",
+               sep = "")
+
+  df <- data.table::fread(cmd = url,
+                          na.strings= NULL,
+                          col.names = c("year", "y"))  %>%
+    as.data.frame %>%
+    dplyr::mutate(state = state) %>%
+    dplyr::select(year, state, y)
 
 }) %>%
   dplyr::bind_rows()