+++ NEWS.md
@@ -1,4 +1,4 @@
 # USgas 0.1.0
 
-* Added the `us_res_gas` dataset, describing the US monthly natural gas residential consumption
+* Added the `us_residential` dataset, describing the US monthly natural gas residential consumption
 * Added a `NEWS.md` file to track changes to the package.